% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_snowwarp_sca.R
\name{calculate_snowwarp_sca}
\alias{calculate_snowwarp_sca}
\title{last modified: Mar 08, 2022}
\usage{
calculate_snowwarp_sca(folder, aoi)
}
\arguments{
\item{folder}{The main directory where the Google Earth Engine files are located on your computer (i.e., the same folder you used for process_snowwarp)}

\item{aoi}{The path for an Area Of Interest shapefile}
}
\value{
The calculate_snowwarp_sca function returns a csv file that indicates for each day in the time series the percentage of the aoi that is covered by snow
}
\description{
This function calculates daily snow covered areas over a specific area of interest using the \code{\link{process_snowwarp}} outputs
}
\examples{
calculate_snowwarp_sca(
folder = 'snowWarpFolder',
aoi = 2004:2008,
outputFolder = 8
)
}
\seealso{
\code{\link{get_snowwarp_tiles}}, \code{\link{extract_snowwarp_stats}}, \code{\link{download_snowwarp_data}}, \code{\link{process_snowwarp}}
}
