% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_snowwarp_tiles.R
\name{get_snowwarp_tiles}
\alias{get_snowwarp_tiles}
\title{Get the number of SnowWarp tiles to be processed}
\usage{
get_snowwarp_tiles(folder)
}
\arguments{
\item{folder}{The main directory where the Google Earth Engine files are located on your computer. There should already be an 'output' folder
inside this directory with the data from the process_snowwarp function.}
}
\value{
The get_num_tiles function returns the number of tiles that need to be processed using the process_snowwarp function.
}
\description{
This function prints a message with the number of SnowWarp tiles to be processed. The number can be used to split up the processing.
}
\examples{
For Mac:
get_snowwarp_tiles(folder = '/Users/MyUser/Documents/GEE_snow_download')

For Windows:
get_snowwarp_tiles(folder = 'D:/GEE_snow_download')

}
\seealso{
\code{\link{process_snowwarp}}, \code{\link{extract_snowwarp_stats}}, \code{\link{download_snowwarp_data}}
}
