% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_snowwarp.R
\name{process_snowwarp}
\alias{process_snowwarp}
\title{Process SnowWarp Data}
\usage{
process_snowwarp(
  folder,
  years,
  cpus,
  file_ext = "",
  tiles = NULL,
  maxmemory = 1e+10
)
}
\arguments{
\item{folder}{The main directory where the Google Earth Engine files are located on your computer}

\item{years}{The years of data you want processed. Must be numeric and between 2000-2018. Snowwarp runs on winter years (Aug 1st - July 31st)
instead of calendar years, so years = 2005 means August 1, 2005 - July 31, 2006, etc.
It is better to run all the years you want at once, and less tiles.}

\item{cpus}{The number of cpus to use for parallel processing. This should not be greater than the number of tiles since each tile is
processed in parallel.}

\item{file_ext}{File name descriptor to add to end of output files. No descriptor by default.}

\item{tiles}{Option to process a certain number of tiles. Run get_snowwarp_tiles first to get the total number of tiles.
It is better to run the function over a smaller number of tiles and all the years desired. Think about ROM and number of processors available.
Defaults to processing all tiles.}

\item{maxmemory}{The amount of memory allocated for the raster package. Defaults to 1e+10, which is slightly more than default.}
}
\value{
The process_snowwarp function returns .tif rasters with daily values of fractional snow covered area. Each file
contains values from August 1st - July 31st of the subsequent year. These daily values can be used
in of themselves, or can be further input into the extract_snowwarp_stats function to calculate annual statistics.
These files are written to 'folder/output'.
}
\description{
This function processes SnowWarp using imagery downloaded from Google Earth Engine. Please refer to instruction manual
\url{https://htmlpreview.github.io/?https://github.com/bermane/snowwarp/blob/master/start_snowwarp.html},
to ensure you have properly downloaded the necessary imagery.
}
\examples{
For Mac:
process_snowwarp(folder = '/Users/MyUser/Documents/GEE_snow_download',
years = 2004:2008,
cpus = 8,
file_ext = 'studyarea1',
tiles = 1:8)

For Windows:
process_snowwarp(folder = 'D:/GEE_snow_download',
years = 2004:2008,
cpus = 8,
file_ext = 'studyarea1',
tiles = 1:8)

}
\seealso{
\code{\link{get_snowwarp_tiles}}, \code{\link{extract_snowwarp_stats}}, \code{\link{download_snowwarp_data}}
}
