% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_snowwarp_data.R
\name{download_snowwarp_data}
\alias{download_snowwarp_data}
\title{Download SnowWarp Data}
\usage{
download_snowwarp_data(
  mail,
  link,
  folder = file.path(getwd(), "snowwarp_data"),
  overwrite = T
)
}
\arguments{
\item{mail}{Your Google Drive account email}

\item{link}{The shareable link to the Google Drive folder that contains the SnowWarp input imagery. (1) Right click on the Google Drive folder.
(2) Obtain shareable link.}

\item{folder}{The folder where you want the imagery to be downloaded to. This will be the same folder you use for the other
SnowWarp functions.}
}
\value{
The download_snowwarp_data function downloads imagery pre-processed on Google Earth Engine. It is necessary to download the imagery
before running the \code{\link{get_snowwarp_tiles}} and \code{\link{process_snowwarp}} functions.
}
\description{
This function downloads imagery from Google Earth Engine that will be input into the SnowWarp algorithm. Please refer to the
instruction manual \url{https://htmlpreview.github.io/?https://github.com/bermane/snowwarp/blob/master/start_snowwarp.html},
and follow the steps to acquire the necessary imagery. You may have issues if Google Chrome is not
your default web browser.
}
\examples{
For Mac:
download_snowwarp_data(mail = 'myemail at server.com',
link = 'shareable google drive folder link',
folder = '/Users/MyUser/Documents/GEE_snow_download')

For Windows:
process_snowwarp(mail = 'myemail at server.com',
link = 'shareable google drive folder link',
folder = 'D:/GEE_snow_download')

}
\seealso{
\code{\link{get_snowwarp_tiles}}, \code{\link{process_snowwarp}}, \code{\link{extract_snowwarp_stats}}
}
