% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_snowwarp_stats.R
\name{extract_snowwarp_stats}
\alias{extract_snowwarp_stats}
\title{Extract SnowWarp Annual Statistics}
\usage{
extract_snowwarp_stats(
  folder,
  cpus,
  file_ext = "",
  tiles = NULL,
  maxmemory = 1e+10
)
}
\arguments{
\item{folder}{The main directory where the Google Earth Engine files are located on your computer. There should already be an 'output' folder
inside this directory with the data from the process_snowwarp function.}

\item{cpus}{The number of cpus to use for parallel processing. This should not be greater than the number of tiles since each tile is
processed in parallel.}

\item{file_ext}{File name descriptor that you added to the process_snowwarp function. No descriptor by default.}

\item{tiles}{Tiles that you have already processed using the process_snowwarp function.}

\item{maxmemory}{The amount of memory allocated for the raster package. Defaults to 1e+10, which is slightly more than default.}
}
\value{
The extract_snowwarp_stats function returns .tif rasters with annual snow statistics (for each winter year). Each file
contains three bands. Band 1 is the date of snow accumulation, Band 2 is the date of snow melt, and Band 3 is the
number of days with snow in the winter year. These files are written to 'folder/output'.
}
\description{
This function extracts SnowWarp annual statistics from the SnowWarp output generated using the function process_snowwarp. The three statistics
calculated are date of snow accumulation, date of snow melt, and number of days with snow in winter year (from Aug 1st - July 31st).
Annual stats will be extracted for all the years of SnowWarp that were processed.
}
\examples{
For Mac:
extract_snowwarp_stats(folder = '/Users/MyUser/Documents/GEE_snow_download',
cpus = 8,
file_ext = 'studyarea1',
tiles = 1:8)

For Windows:
extract_snowwarp_stats(folder = 'D:/GEE_snow_download',
cpus = 8,
file_ext = 'studyarea1',
tiles = 1:8)

}
\seealso{
\code{\link{process_snowwarp}}, \code{\link{get_snowwarp_tiles}}, \code{\link{download_snowwarp_data}}
}
